# src/openapi/schema.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EduCRM API",
    "version": "1.0.0",
    "description": "Educational CRM + LMS Platform REST API — Full Schema"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["auth"],
        "operationId": "authLogin",
        "summary": "Login with email and password",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthTokensResponse" }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email":      { "type": "string", "format": "email", "example": "admin@example.com" },
          "password":   { "type": "string", "minLength": 8, "example": "password123" },
          "tenantSlug": { "type": "string", "example": "academy1" }
        }
      },
...
      "AuthTokensResponse": {
        "type": "object",
        "required": ["accessToken", "refreshToken", "expiresIn"],
        "properties": {
          "accessToken":  { "type": "string" },
          "refreshToken": { "type": "string" },
          "expiresIn":    { "type": "integer", "description": "Token lifetime in seconds", "example": 900 }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["statusCode", "message", "code"],
        "properties": {
          "statusCode": { "type": "integer" },
          "message":    { "type": "string" },
          "code":       { "type": "string" }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "required": ["statusCode", "message", "code", "errors"],
        "properties": {
          "statusCode": { "type": "integer", "example": 422 },
          "message":    { "type": "string", "example": "Validation failed" },
          "code":       { "type": "string", "example": "VALIDATION_ERROR" },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": { "type": "string" }
            },
            "example": {
              "email": ["Must be a valid email address"],
              "password": ["Must be at least 8 characters"]
            }
          }
        }
      }
    }
  }
}